bugfix in offsetX constraint
wrong usage of abs() instead of max() function
minor improvements

+++ SmartMeter/GraphViewController.swift
@@ -56,7 +56,7 @@ class GraphViewController: UIViewController {
             if powerGraphEngine != nil {
                 let translation = gesture.translationInView(graphView)
                 let samplesPerPoint = CGFloat(powerGraphEngine!.graphViewgetSampleCount()) / graphView.bounds.width
-                powerGraphEngine?.offsetX -= Int(translation.x * samplesPerPoint)
+                powerGraphEngine?.offsetX -= Double(translation.x * samplesPerPoint)
                 gesture.setTranslation(CGPointZero, inView: graphView)
                 graphView.setNeedsDisplay()
             }
